requirements.txt
zappa==0.58.0
urllib3==1.26.18
pillow==10.0.1
boto3==1.28.72 